Refactor: Center BottomAppBar icons on Devocionales page
This commit refactors the `BottomAppBar` on the `DevocionalesPage` to ensure its icons are properly centered.

The `Row` containing the `IconButton`s has been wrapped in a `Center` widget, and the `padding` of the `BottomAppBar` has been set to zero. The height of the bar was also slightly increased from 56 to 60.